flask_blog2/requirements.txt
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
pip==23.1.2
setuptools==65.5.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.0
Pillow==10.1.0